requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=4.0.0
django-filter<25
drf-spectacular>=0.26.0
psycopg2-binary>=2.9.5
python-decouple>=3.8
dj-database-url>=2.0.0
django-redis>=5.2.0
redis>=4.5.0
celery>=5.2.7
gunicorn>=21.2.0
boto3>=1.26.0
django-storages>=1.13.0
Pillow>=10.0.0
stripe>=5.4.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
reportlab>=4.0.0
openpyxl>=3.1.0

# Push Notifications
fcm-django>=2.0.0
firebase-admin>=6.5.0

# Testing dependencies
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-mock>=3.11.1
factory-boy>=3.3.0
